updated redshift spec and code example
