+++ javav2/example_code/redshift/src/main/java/com/example/redshift/User.java
@@ -5,13 +5,14 @@ package com.example.redshift;
 public class User {
 
     private String username;
-    private String password;
+    private String  password;
 
     private String masterUsername;
 
     private String masterUserPassword;
 
     private String host;
+
 
     //getter
     String getUsername(){
